deleted search bar component and added it into our main episode page due to a bug
